+++ Arma3ObjectBuilder/io/export_p3d.py
@@ -151,8 +151,7 @@ def merge_proxy_objects(main_obj, proxy_objects, relative):
         utils.create_selection(proxy, placeholder)
         proxy_lookup[placeholder] = proxy.a3ob_properties_object_proxy.to_placeholder(relative)
 
-        for uv in proxy.data.uv_layers:
-            proxy.data.uv_layers.remove(uv)
+        utils.clear_uvs(proxy)
 
     all_objects = proxy_objects + [main_obj]
     for obj in proxy_objects:
@@ -293,8 +292,7 @@ def get_lod_data(operator, context, validator, temp_collection):
             bm.free()
 
         if main_obj.a3ob_properties_object.lod not in allow_uv:
-            for uv in main_obj.data.uv_layers:
-                main_obj.data.uv_layers.remove(uv)
+            utils.clear_uvs(main_obj)
 
         lod_list.append((main_obj, proxy_lookup, is_valid))